+++ packages/expo/CHANGELOG.md
@@ -14,6 +14,7 @@
 - [cli] Added modules for creating dev server URLs, akin to `UrlUtils` in `xdl`. ([#16557](https://github.com/expo/expo/pull/16557) by [@EvanBacon](https://github.com/EvanBacon))
 - [cli] Added shim for `expo start` command and option resolvers. ([#16587](https://github.com/expo/expo/pull/16587) by [@EvanBacon](https://github.com/EvanBacon))
 - [cli] Added modules for interacting with Metro and Webpack bundlers. ([#16517](https://github.com/expo/expo/pull/16517) by [@EvanBacon](https://github.com/EvanBacon))
+- [cli] Added module for interacting with Metro bundler. ([#16631](https://github.com/expo/expo/pull/16631) by [@EvanBacon](https://github.com/EvanBacon))
 
 ### 🐛 Bug fixes